Gradienty naprawione, błąd był przy obliczaniu splotu (konwolucji), było "<" zamiast "<=". Poza tym jakieś komentarze itd.

+++ SysWizyjne/src/enums/GradientMask.java
@@ -7,21 +7,21 @@ package enums;
  */
 public enum GradientMask 
 {
-	ROBERTSX(new int[][]{{0,0,0},{-1,0,0},{0,1,0}}),
-	ROBERTSY(new int[][]{{0,0,0},{0,0,-1},{0,1,0}}),
-	SOBELX(new int[][]{{-1,-2,-1},{0,0,0},{1,2,1}}),
-	SOBELY(new int[][]{{-1,0,1},{-2,0,2},{-1,0,1}}),
-	PREWITTX(new int[][]{{-1,-1,-1},{0,0,0},{1,1,1}}),
-	PREWITTY(new int[][]{{-1,0,1},{-1,0,1},{-1,0,1}});
+	ROBERTSX(new double[][]{{0,0,0},{-1,0,0},{0,1,0}}),
+	ROBERTSY(new double[][]{{0,0,0},{0,0,-1},{0,1,0}}),
+	SOBELX(new double[][]{{-1,-2,-1},{0,0,0},{1,2,1}}),
+	SOBELY(new double[][]{{-1,0,1},{-2,0,2},{-1,0,1}}),
+	PREWITTX(new double[][]{{-1,-1,-1},{0,0,0},{1,1,1}}),
+	PREWITTY(new double[][]{{-1,0,1},{-1,0,1},{-1,0,1}});
 	
-	private final int[][] _matrix;
+	private final double[][] _matrix;
 	
-	GradientMask(int[][] matrix)
+	GradientMask(double[][] matrix)
 	{
 		_matrix=matrix;
 	}
 	
-	public int[][] getMask()
+	public double[][] getMask()
 	{
 		return _matrix;
 	}
